# api.yml
...
swagger: '2.0'
info:
  version: 1.0.0
  title: api-activos-fijo Microservice
  description: Java Microservice + MySQL Database
  contact:
    name: Jhon alexander chacon torres
    email: jhachacont@correo.udistrital.edu.co
  license:
    name: PruebaConocimiento
host: localhost
basePath: /api-activos-fijo-ms/resources
schemes:
- http
consumes:
- application/json
produces:
- application/json
paths:
...
  /areas:
    get:
      summary: retorna todos los registros de areas
      operationId: getAllAreas
      parameters:
      - name: firstResult
        in: query
        description: el primer resultado a retornar
        required: false
        type: integer
      - name: maxResult
        in: query
        description: el maximo numero de resultados a retornar
        required: false
        type: integer
      responses:
        200:
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/Areas'
        406:
          description: Not Acceptable (Invalid Params)
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
    post:
      summary: crear nuevas areas
      operationId: createAreas
      parameters:
      - name: areas
        in: body
        required: true
        description: Objeto Areas a crear
        schema:
          $ref: '#/definitions/Areas'
      responses:
        201:
          description: Respuesta para la creacion del objeto
          schema:
            $ref: '#/definitions/Areas'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        403:
          description: Forbidden el dato ya existe
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
    put:
      summary: Actualizar un objeto de areas
      operationId: updateAreas
      parameters:
      - name: areas
        in: body
        required: true
        description: Objeto areas a actualizar
        schema:
          $ref: '#/definitions/Areas'
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/Areas'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        404:
          description: Not Found no existe el dato
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'
    delete:
      summary: Eliminar areas segun llave o codigo
      operationId: deleteAreas
      parameters:
      - name: codigo
        in: query
        required: true
        description: codigo o lleve del area
        type: string
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/Areas'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        404:
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'
  /areas/byCodigo:
    get:
      summary: Retorna un area segun su codigo
      operationId: getAreaByCode
      parameters:
      - name: codigo
        in: query
        description: codigo del area
        required: true
        type: string
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/Areas'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        404:
          description: Not Found no existe el dato
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'
...
  /asociacionPersonas:
    get:
      summary: retorna todos los registros de la asociacion de personas
      operationId: getAllAsociacionPersonas
      parameters:
      - name: firstResult
        in: query
        description: el primer resultado a retornar
        required: false
        type: integer
      - name: maxResult
        in: query
        description: el maximo numero de resultados a retornar
        required: false
        type: integer
      responses:
        200:
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/AsociacionPersonas'
        406:
          description: Not Acceptable (Invalid Params)
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
    post:
      summary: crear nuevas asociaciones de personas
      operationId: createAsociacionPersonas
      parameters:
      - name: asociacionPersonas
        in: body
        required: true
        description: Objeto AsociacionPersonas a crear
        schema:
          $ref: '#/definitions/AsociacionPersonas'
      responses:
        201:
          description: Respuesta para la creacion del objeto
          schema:
            $ref: '#/definitions/AsociacionPersonas'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        403:
          description: Forbidden el dato ya existe
          schema:
            $ref: '#/definitions/ErrorModel'
        406:
          description: Not Acceptable (Invalid Params)
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
    put:
      summary: Actualizar un objeto de asociacion personas
      operationId: updateAsociacionPersonas
      parameters:
      - name: asociacionPersonas
        in: body
        required: true
        description: Objeto asociacionPersonas a actualizar
        schema:
          $ref: '#/definitions/AsociacionPersonas'
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/AsociacionPersonas'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        404:
          description: Not Found no existe el dato
          schema:
            $ref: '#/definitions/ErrorModel'
        406:
          description: Not Acceptable (Invalid Params)
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'
    delete:
      summary: Eliminar asociacion personas segun llave o codigo
      operationId: deleteAsociacionPersonas
      parameters:
      - name: codigo
        in: query
        required: true
        description: codigo o lleve de la asociacion de la persona
        type: number
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/AsociacionPersonas'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        404:
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'
  /asociacionPersonas/byCodigo:
    get:
      summary: Retorna una asociacion de persona segun su codigo
      operationId: getAsociacionPersonasByCode
      parameters:
      - name: codigo
        in: query
        description: codigo de la asociacion de persona
        required: true
        type: number
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/AsociacionPersonas'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        404:
          description: Not Found no existe el dato
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'
definitions:
  Tipo:
    type: object
    required: ["codigo", "descripcion"]
    properties:
      codigo:
        type: string
      descripcion:
        type: string
  Color:
    type: object
    required: ["codigo", "descripcion"]
    properties:
      codigo:
        type: string
      descripcion:
        type: string
  EstadoActual:
    type: object
    required: ["codigo", "descripcion"]
    properties:
      codigo:
        type: string
      descripcion:
        type: string
  Personas:
    type: object
    required: ["documento", "nombres", "apellidos", "telefono", "correo", "direccion"]
    properties:
      documento:
        type: string
      nombres:
        type: string
      apellidos:
        type: string
      telefono:
        type: string
      correo:
        type: string
      direccion:
        type: string
  Areas:
    type: object
    required: ["documento", "nombre", "ciudad", "descripcion"]
    properties:
      documento:
        type: string
      nombre:
        type: string
      ciudad:
        type: string
      descripcion:
        type: string
  Activos:
    type: object
    required: ["codigo", "nombre", "descripcion", "tipo", "numeroInternoInventario", "peso", "alto", "largo", "serial", "valorCompra", "fechaCompra", "fechaBaja", "estadoActual", "color"]
    properties:
      codigo:
        type: string
      nombre:
        type: string
      descripcion:
        type: string
      tipo:
          $ref: '#/definitions/Tipo'
      numeroInternoInventario:
        type: string
      peso:
        type: number
      alto:
        type: number
      largo:
        type: number
      serial:
        type: string
      valorCompra:
        type: number
      fechaCompra:
        type: number
      fechaBaja:
        type: number
      estadoActual:
        $ref: '#/definitions/EstadoActual'
      color:
        $ref: '#/definitions/Color'
...
  AsociacionPersonas:
    type: object
    required: ["codigo", "personas", "activos", "fechaAsignacion", "fechaRetiro", "descripcion"]
    properties:
      codigo:
        type: number
      personas:
        $ref: '#/definitions/Personas'
      activos:
        $ref: '#/definitions/Activos'
      fechaAsignacion:
        type: number
      fechaRetiro:
        type: number
      descripcion:
        type: string


  # Errors
  ErrorModel:
    type: object
    required:
    - code
    - description
    - id
    properties:
      code:
        type: integer
      description:
        type: object
        properties:
          key:
            type: string
          value:
            type: string
      id:
        type: string